+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Backupcenter/Search-AzDataProtectionBackupInstanceInAzGraph.ps1
@@ -1,4 +1,4 @@
-﻿function Search-AzDataProtectionBackupInstanceInAzGraph
+function Search-AzDataProtectionBackupInstanceInAzGraph
 {
 	[OutputType('PSObject')]
     [CmdletBinding(PositionalBinding=$false)]