+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Common/New-AzDataProtectionResourceGuard.ps1
@@ -1,4 +1,4 @@
-﻿
+
 
 function New-AzDataProtectionResourceGuard
 {   
@@ -27,6 +27,10 @@ function New-AzDataProtectionResourceGuard
         [Parameter(ParameterSetName="CreateResourceGuard", Mandatory=$false, HelpMessage='Optional ETag')]
         [System.String]
         ${ETag},
+
+        [Parameter(ParameterSetName="CreateResourceGuard", Mandatory=$false, HelpMessage='This parameter is no longer in use and will be deprecated')]
+        [System.String]
+        ${IdentityType},
                 
         [Parameter(ParameterSetName="CreateResourceGuard", Mandatory, HelpMessage='Location of the resource guard')]
         [System.String]
@@ -79,6 +83,13 @@ function New-AzDataProtectionResourceGuard
 
     process
     {
+        if($PSBoundParameters.ContainsKey("IdentityType"))
+        {
+            $null = $PSBoundParameters.Remove("IdentityType")
+
+            # TODO : need to move this to parameter level 
+            Write-Warning "Parameter IdentityType is no longer in use and will be deprecated in upcoming breaking change release"
+        }
         Az.DataProtection.Internal\New-AzDataProtectionResourceGuard @PSBoundParameters
     }
 }