+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Common/Update-AzDataProtectionResourceGuard.ps1
@@ -1,4 +1,4 @@
-﻿
+
 
 function Update-AzDataProtectionResourceGuard
 {   
@@ -27,6 +27,10 @@ function Update-AzDataProtectionResourceGuard
         [Parameter(ParameterSetName="UpdateResourceGuardOperations", Mandatory=$false, HelpMessage='Optional ETag')]
         [System.String]
         ${ETag},
+
+        [Parameter(ParameterSetName="UpdateResourceGuardOperations", Mandatory=$false, HelpMessage='This parameter is no longer in use and will be deprecated')]
+        [System.String]
+        ${IdentityType},
         
         [Parameter(ParameterSetName="UpdateResourceGuardOperations", Mandatory=$false, HelpMessage='Resource tags')]        
         [Hashtable]
@@ -110,6 +114,13 @@ function Update-AzDataProtectionResourceGuard
 
         # Add Location
         $null = $PSBoundParameters.Add("Location", $ResGuard.Location)
+        
+        if($PSBoundParameters.ContainsKey("IdentityType"))
+        {
+            $null = $PSBoundParameters.Remove("IdentityType")
+            # TODO : need to move this to parameter level 
+            Write-Warning "Parameter IdentityType is no longer in use and will be deprecated in upcoming breaking change release"
+        }
 
         Az.DataProtection.Internal\New-AzDataProtectionResourceGuard @PSBoundParameters
     }