+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/BackupInstance/Initialize-AzDataProtectionBackupInstance.ps1
@@ -1,8 +1,7 @@
-﻿function Initialize-AzDataProtectionBackupInstance {
+function Initialize-AzDataProtectionBackupInstance {
     [OutputType('Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Models.IBackupInstanceResource')]
     [CmdletBinding(PositionalBinding=$false)]
     [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Description('Initializes Backup instance Request object for configuring backup')]
-    [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.ModelCmdletAttribute()]
 
     param(
         [Parameter(Mandatory=$false, HelpMessage='Policy Id to be associated to Datasource')]