+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/BackupInstance/Suspend-AzDataProtectionBackupInstanceBackup.ps1
@@ -1,4 +1,4 @@
-﻿
+
 
 function Suspend-AzDataProtectionBackupInstanceBackup
 {