+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/Restore/Start-AzDataProtectionBackupInstanceRestore.ps1
@@ -1,4 +1,4 @@
-﻿
+
 
 function Start-AzDataProtectionBackupInstanceRestore
 {   
@@ -54,7 +54,7 @@ function Start-AzDataProtectionBackupInstanceRestore
         ${RestoreTargetInfo},
 
         [Parameter(ParameterSetName="TriggerExpanded", Mandatory, HelpMessage='Type of the source data store')]
-        [System.String]
+        [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Support.SourceDataStoreType]
         ${SourceDataStoreType},
 
         [Parameter(ParameterSetName="TriggerExpanded", Mandatory=$false, HelpMessage='ARM URL for User Assigned Identity')]