+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Platform/Restore/Test-AzDataProtectionBackupInstanceRestore.ps1
@@ -1,4 +1,4 @@
-﻿
+
 
 function Test-AzDataProtectionBackupInstanceRestore
 {