+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Policy/Edit-AzDataProtectionPolicyRetentionRuleClientObject.ps1
@@ -1,4 +1,4 @@
-﻿
+
 
 function Edit-AzDataProtectionPolicyRetentionRuleClientObject {
     [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.ModelCmdletAttribute()]