+++ src/DataProtection/DataProtection.Autorest/custom/Cmdlets/Policy/Get-AzDataProtectionPolicyTemplate.ps1
@@ -1,4 +1,4 @@
-﻿
+
 
 function Get-AzDataProtectionPolicyTemplate {
     [Microsoft.Azure.PowerShell.Cmdlets.DataProtection.ModelCmdletAttribute()]