+++ src/DataProtection/DataProtection.Autorest/custom/Helpers/CommonHelpers.ps1
@@ -1,4 +1,4 @@
-﻿
+
 function GetDatasourceSetInfo
 {
 	[Microsoft.Azure.PowerShell.Cmdlets.DataProtection.DoNotExportAttribute()]