+++ src/DataProtection/DataProtection.Autorest/custom/Models/BackupInstanceResource.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
@@ -9,9 +9,9 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Models
         private string backupInstanceName;
         public string BackupInstanceName
         {
-            get
+            get 
             {
-                if (this.Name != null && this.Name != "")
+                if(this.Name != null && this.Name != "")
                 {
                     return this.Name;
                 }