+++ src/DataProtection/DataProtection.Autorest/custom/Models/DeletedBackupVaultResource.cs
@@ -10,14 +10,14 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Models
         private string location;
         public string Location
         {
-            get
+            get 
             {
                 // Extract location from Id
                 // Format: /subscriptions/{subscriptionId}/providers/Microsoft.DataProtection/locations/{location}/deletedVaults/{vaultId}
-                if (string.IsNullOrEmpty(this.location) && !string.IsNullOrEmpty(this.Id))
+                if(string.IsNullOrEmpty(this.location) && !string.IsNullOrEmpty(this.Id))
                 {
                     var match = Regex.Match(this.Id, @"/locations/([^/]+)/", RegexOptions.IgnoreCase);
-                    if (match.Success)
+                    if(match.Success)
                     {
                         this.location = match.Groups[1].Value;
                     }
@@ -30,14 +30,14 @@ namespace Microsoft.Azure.PowerShell.Cmdlets.DataProtection.Models
         private string originalBackupVaultResourceGroup;
         public string OriginalBackupVaultResourceGroup
         {
-            get
+            get 
             {
                 // Extract resource group from OriginalBackupVaultResourcePath
                 // Format: /subscriptions/{subscriptionId}/resourcegroups/{resourceGroupName}/providers/Microsoft.DataProtection/BackupVaults/{vaultName}
-                if (string.IsNullOrEmpty(this.originalBackupVaultResourceGroup) && !string.IsNullOrEmpty(this.OriginalBackupVaultResourcePath))
+                if(string.IsNullOrEmpty(this.originalBackupVaultResourceGroup) && !string.IsNullOrEmpty(this.OriginalBackupVaultResourcePath))
                 {
                     var match = Regex.Match(this.OriginalBackupVaultResourcePath, @"/resourcegroups/([^/]+)/", RegexOptions.IgnoreCase);
-                    if (match.Success)
+                    if(match.Success)
                     {
                         this.originalBackupVaultResourceGroup = match.Groups[1].Value;
                     }